moved CancelSignUpButton to its own separate component file
