+++ src/components/UpcomingEventsList/index.tsx
@@ -1,52 +1,13 @@
 'use client';
 import React from 'react';
 import { EventEntity } from '@/types/dataModel/event';
-import {
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContentText,
-  List,
-  ListItem,
-} from '@mui/material';
+import { Box, List, ListItem } from '@mui/material';
+import CancelSignUpButton from '../CancelSignUpButton';
 
 interface EventObjectList {
   events: EventEntity[];
   volunteerID: string;
 }
-
-const CancelSignUp = (props: { event: EventEntity }) => {
-  const [open, setOpen] = React.useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-    console.log(props.event.title);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <>
-      <Button variant="contained" onClick={handleClickOpen}>
-        Cancel Sign up
-      </Button>
-      <Dialog onClose={handleClose} open={open}>
-        <DialogContentText sx={{ margin: '15px' }}>
-          You have cancelled your sign-up for event &quot;{props.event.title}
-          &quot;
-        </DialogContentText>
-        <DialogActions>
-          <Button onClick={handleClose} autoFocus>
-            OK
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </>
-  );
-};
 
 export default function UpcomingEventsList(props: EventObjectList) {
   const isFutureEvent = (eventDate: Date) => {
@@ -70,7 +31,7 @@ export default function UpcomingEventsList(props: EventObjectList) {
           >
             {event.title}
             {isFutureEvent(event.day) && (
-              <CancelSignUp event={event}></CancelSignUp>
+              <CancelSignUpButton event={event}></CancelSignUpButton>
             )}
           </ListItem>
         ))}
